All HTML, CSS, JS, and PHP files for the site
These pages are not as organized as they could be due to the lack of consistency in which I have updated them. It works for what my website needs but does not serve as a good example of quality code. Most of the php and javascript pages are the first time I have explored these languages. I'm constantly updating and learning!